fix(check): mirror absolute project imports
